services/route_delivery/requirements.txt
nats-py>=2.7.0
asyncpg>=0.29.0
httpx>=0.27.0
paho-mqtt>=1.6.1,<2.0

